backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.6
pytest==8.2.0
python-dotenv==1.0.0
redis==4.5.5
requests==2.31.0
rq==1.10
Werkzeug==2.3.7

